Move messages from IdCardConsoleComponent.cs to SharedIdCardSystem.cs
