+++ Content.Shared/Access/Components/IdCardConsoleComponent.cs
@@ -22,38 +22,6 @@ public sealed partial class IdCardConsoleComponent : Component
     [DataField]
     public ItemSlot TargetIdSlot = new();
 
-    [Serializable, NetSerializable]
-    public sealed class WriteToTargetIdMessage : BoundUserInterfaceMessage
-    {
-        public readonly string FullName;
-        public readonly string JobTitle;
-        public readonly List<string> AccessList;
-        public readonly string JobPrototype;
-
-        public WriteToTargetIdMessage(string fullName, string jobTitle, List<string> accessList, string jobPrototype)
-        {
-            FullName = fullName;
-            JobTitle = jobTitle;
-            AccessList = accessList;
-            JobPrototype = jobPrototype;
-        }
-    }
-
-    [Serializable, NetSerializable]
-    public sealed class WriteToShuttleDeedMessage : BoundUserInterfaceMessage
-    {
-        public readonly string ShuttlePrefix;
-        public readonly string ShuttleName;
-        public readonly string ShuttleSuffix;
-
-        public WriteToShuttleDeedMessage(string shuttlePrefix, string shuttleName, string shuttleSuffix)
-        {
-            ShuttlePrefix = shuttlePrefix;
-            ShuttleName = shuttleName;
-            ShuttleSuffix = shuttleSuffix;
-        }
-    }
-
     // Put this on shared so we just send the state once in PVS range rather than every time the UI updates.
 
     [DataField, AutoNetworkedField]
